Force input type to be an email
When viewing the login form on a mobile device, there are variations of
keyboards that are presented to the user depening what input type has
been set on the field.

Our testers have found it less usable to enter an email address, due to
need of switching between alpha to numeric characters. Lack of an @ is
also not ideal.

We've had two choices, either, set the `inputmode="email"` or
`type="email"`. The difference between the two, is additional validation
done by the browser of the `type` attribute. Since we have disabled the
validation on the form entirely, and we don't care for the browser
validation, `type` attribute is less "effort" to implement.